Restructure admin panel into tabbed layout
Extract appearance settings into dedicated component and add
Skeleton UI Tabs with Appearance, Workflows, and Appointments tabs.
Workflows and Appointments are placeholders for #63 and #64.

Closes #62
