+++ client/src/lib/components/admin/AppointmentsTab.svelte
@@ -1,0 +1,9 @@
+<div class="space-y-8">
+	<section class="card p-6 space-y-4">
+		<h2 class="h3">Appointments</h2>
+		<p class="opacity-75">
+			Manage your available time slots for phone calls and meetings via Google Calendar.
+		</p>
+		<p class="text-sm opacity-50">Coming soon.</p>
+	</section>
+</div>
