+++ client/src/lib/components/admin/WorkflowsTab.svelte
@@ -1,0 +1,9 @@
+<div class="space-y-8">
+	<section class="card p-6 space-y-4">
+		<h2 class="h3">Workflows</h2>
+		<p class="opacity-75">
+			Configure the steps visitors must complete to access your contacts or agenda.
+		</p>
+		<p class="text-sm opacity-50">Coming soon.</p>
+	</section>
+</div>
